перенес логику рендеринга панели в контроллер
